works/requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyMuPDF==1.19.6
python-docx==0.8.11
vercel==0.2.1